Add various tools
### This time I added:
new category "Web Proxy and Traffic Interception", In which I added tools:
- mitmproxy
- proxify
- FoxyProxy Browser Extension
- zaproxy

for the last one "zapproxy" which is already in "Vulnerability Scanners" category, still I added it here because it's a scanner and proxy tool both, just like "Httpx" is in "Technologies" category and in "Uncategorized" also because it do more that just detecting tech so it need to be both place. By at this example i tried to put "zapproxy" in "Web Proxy and Traffic Interception" category.
I explained the reasoning, if want any changes, let me know.

### Other tools I added:
- wafw00f
- nuclei-templates (I added it right next to nuclei tool I thought these two repositories are related and placing them next to each other might help people find the connection more easily. ) (I hope it's fine)
- notify
- interactsh
- asnmap
- mapcidr
- tlsx
- CSPBypass
- Big Bounty Recon
- Bypass Bot Detection

### Other minor changes
1. I removed an empty line "524"
2. The tool "cariddi" was in "Vulnerability Scanners" category but it's also a tool which looks for secrets, api tokens, etc. So I put it in "Secrets" category also. (Just like what i did with "zapproxy")

Let me know if there's any changes to made!